+++ src/rdfcrate/wrapper.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 from pathlib import Path
 from typing import Annotated, Any, Iterable, TypeVar, TYPE_CHECKING
-from typing_extensions import Doc
-from rdflib import RDF, Graph, URIRef
-from rdfcrate.vocabs import rdf
-from rdfcrate.rdfprop import RdfProperty, ReverseProperty
-from rdfcrate.rdfterm import RdfTerm
-from rdfcrate.rdftype import RdfClass, EntityArgs
+from deprecated import deprecated
+from typing_extensions import Doc, Unpack
+from rdflib import URIRef
+from rdfcrate.context_graph import ContextGraph, EntityArgs, ContextGraphKwargs
+from rdfcrate.rdfprop import RdfProperty
+from rdfcrate.rdftype import RdfClass
 from rdfcrate.spec_version import SpecVersion, ROCrate1_1
-from dataclasses import InitVar, dataclass, field
+from dataclasses import dataclass
 import mimetypes
 from os import stat
 from abc import ABCMeta, abstractmethod
@@ -42,23 +42,33 @@ def has_prop(props: Iterable[EntityArgs], predicate: type[RdfProperty]) -> bool:
     return any(isinstance(prop, predicate) for prop in props)
 
 
-@dataclass
-class RoCrate(metaclass=ABCMeta):
+class RoCrateKwargs(ContextGraphKwargs, total=False):
+    version: SpecVersion
+
+
+class RoCrate(ContextGraph, metaclass=ABCMeta):
     """
     Abstract class containing common functionality for both attached and detached RO-Crates
     """
 
-    graph: Graph = field(init=False, default_factory=Graph)
-    "`rdflib.Graph` containing the RO-Crate metadata. This can be accessed directly, but it is recommended to use the other methods provided by this class where available."
-    custom_terms: dict[str, str] = field(init=False, default_factory=dict)
-    "Set of custom terms not in the standard RO-Crate context that need to be in the final JSON-LD context."
-    context: Context = field(init=False)
-    "Dynamically updated version of the JSON-LD context. By default, this contains the current RO-Crate spec's context, but additional terms are added as they are added to the graph."
-    version: SpecVersion = field(kw_only=True, default=ROCrate1_1)
+    version: SpecVersion
     "Version of the RO-Crate specification to use"
+    _custom_terms: dict[str, Any]
 
-    def __post_init__(self):
-        self.context = Context(self.version.context_url)
+    def __init__(self, **kwargs: Unpack[RoCrateKwargs]):
+        super().__init__(**kwargs)
+        self.version = kwargs.get("version", ROCrate1_1)
+
+        # We resolve the RO-Crate context now so we can use it to resolve terms and check if they are already registered
+        self._roc_context = Context(self.version.context_url)
+
+    @property
+    def context_source(self):
+        # Implements the requirements for context format according to the RO-Crate spec
+        if len(self._custom_terms) == 0:
+            return self.version.context_url
+        else:
+            return [self.version.context_url, self._custom_terms]
 
     @property
     @abstractmethod
@@ -67,67 +77,7 @@ class RoCrate(metaclass=ABCMeta):
         The root entity of the RO-Crate
         """
 
-    def register_terms(self, terms: Iterable[RdfTerm]) -> None:
-        """
-        Adds custom terms to the crate context
-        """
-        for term in terms:
-            existing = self.context.expand(term.label)
-            if existing == str(term.uri):
-                # Skip terms that are already in the RO-Crate context
-                continue
-            if existing is not None:
-                raise ValueError(
-                    f'Term "{term.label}" is already defined to mean {existing}. Cannot redefine to {term.uri}.'
-                )
-
-            if term.uri == RDF.type:
-                # rdf:type should never be re-defined
-                continue
-
-            # The context keeps track of all terms, including the base RO-Crate terms.
-            self.context.add_term(term.label, str(term.uri))
-            # Custom terms only tracks the non-standard terms
-            self.custom_terms[term.label] = str(term.uri)
-
-    def add_entity(self, entity: EntityClass, *args: EntityArgs) -> EntityClass:
-        """
-        Adds any type of entity to the crate
-
-        Returns:
-            The ID of the new entity
-
-        Example:
-            ```python
-            from rdflib import BNode,
-            from rdfcrate.vocabs import schemaorg
-
-            crate.add_entity(
-                BNode(),
-                schemaorg.Person,
-                schemaorg.name("Alice")
-            )
-            ```
-        """
-
-        self.register_terms(
-            # Register property terms
-            [prop.term for prop in args]
-            +
-            # Register class terms
-            [entity.term]
-            # Register supplementary class terms
-            + [
-                prop.object.term
-                for prop in args
-                # rdf:type is a special case. All other objects do not need to be registered
-                # since they are allowed to be full IRIs
-                if isinstance(prop, RdfProperty) and isinstance(prop, rdf.type)
-            ]
-        )
-        entity.add(self.graph, *args)
-        return entity
-
+    @deprecated("This is deprecated in favour of `.add(RootDataEntity(...))`")
     def add_root_entity(
         self,
         name: schemaorg.name,
@@ -155,6 +105,7 @@ class RoCrate(metaclass=ABCMeta):
         The metadata entity of the RO-Crate
         """
 
+    @deprecated("This is deprecated in favour of `.add(MetadataEntity(...))`")
     def add_metadata_entity(self, *props: EntityArgs) -> None:
         """
         Creates the `ro-crate-metadata.json` record.
@@ -184,6 +135,7 @@ class RoCrate(metaclass=ABCMeta):
             dataset = self.root_data_entity
         self.add_metadata(dataset, schemaorg.hasPart(entity))
 
+    @deprecated("This is deprecated in favour of `.add(FileEntity(...))`")
     def register_file(
         self,
         path: str,
@@ -228,6 +180,7 @@ class RoCrate(metaclass=ABCMeta):
 
         return file_id
 
+    @deprecated("This is deprecated in favour of `.add(DirEntity(...))`")
     def register_dir(
         self,
         path: str,
@@ -265,31 +218,7 @@ class RoCrate(metaclass=ABCMeta):
             self.link_to_dataset(dir_id, dataset)
         return dir_id
 
-    def add_metadata(self, entity: RdfClass, *args: EntityArgs) -> None:
-        """
-        Add metadata for an existing entity.
-
-        Returns:
-            The ID of the updated entity
-
-        Params:
-            uri: ID of the entity being described
-        """
-        self.register_terms(
-            # Register property terms
-            [prop.term for prop in args]
-            + 
-            # Register supplementary class terms
-            [
-                prop.object.term
-                for prop in args
-                if isinstance(prop, RdfProperty) and isinstance(prop, rdf.type)
-            ]
-        )
-        for arg in args:
-            arg.add_to_graph(self.graph, entity.id)
-
-    def compile(self) -> str:
+    def compile(self, **kwargs) -> str:
         """
         Compiles the RO-Crate to a JSON-LD string
         """
@@ -298,30 +227,20 @@ class RoCrate(metaclass=ABCMeta):
                 "Root data entity not found. Did you call `add_root_entity`?"
             )
 
-        if len(self.custom_terms) == 0:
-            context = self.version.context_url
-        else:
-            context = [self.version.context_url, self.custom_terms]
-
-        # Serializer kwargs are annoyingly not listed in the docs.
-        # See them here: https://github.com/RDFLib/rdflib/blob/d220ee3bcba10a7af6630c4faaa37ca9cee33554/rdflib/plugins/serializers/jsonld.py#L76-L84
-        return self.graph.serialize(format="json-ld", context=context)
+        return super().compile(**kwargs)
 
 
-@dataclass
 class AttachedCrate(RoCrate):
     """
     See <https://www.researchobject.org/ro-crate/specification/1.2-DRAFT/structure#attached-ro-crate>
     """
 
-    path: InitVar[str | Path]
-    "The RO-Crate directory"
-    root: Path = field(init=False)
+    root: Path
     "The RO-Crate directory"
 
-    def __post_init__(self, path: str | Path):
+    def __init__(self, path: str | Path, **kwargs: Unpack[RoCrateKwargs]):
+        super().__init__(**kwargs)
         self.root = Path(path).resolve()
-        super().__post_init__()
 
     @property
     def metadata_entity(self) -> schemaorg.CreativeWork:
@@ -413,7 +332,7 @@ class AttachedCrate(RoCrate):
         description: schemaorg.description,
         date_published: schemaorg.datePublished,
         license: schemaorg.license,
-        *props: RdfProperty | ReverseProperty,
+        *props: EntityArgs,
         recursive: Recursive = False,
     ) -> schemaorg.Dataset:
         return self.register_dir(