+++ test/test_crate.py
@@ -1,8 +1,7 @@
 from pathlib import Path
-from rdfcrate import AttachedCrate, RdfProperty, owl, RdfClass
-from rdflib import RDF, Literal, Graph, BNode, URIRef
-from rdfcrate.rdfterm import RdfTerm
-from rdfcrate.vocabs import dc, sdo, roc, rdf, bioschemas_drafts, rdfs
+from rdfcrate import AttachedCrate
+from rdflib import Literal, Graph, URIRef
+from rdfcrate.vocabs import dc, sdo, roc, rdf
 import json
 from datetime import datetime
 import tempfile
@@ -20,7 +19,7 @@ BASE_SUBJECTS = [MIT, ME, WEHI_RCP]
 @pytest.fixture()
 def empty_crate():
     with tempfile.TemporaryDirectory() as tmpdir:
-        yield AttachedCrate(tmpdir)
+        yield AttachedCrate(path=tmpdir)
 
 @pytest.fixture(scope="function", autouse=True)
 def rocrate_context():
@@ -137,143 +136,6 @@ def test_mime_type():
         URIRef("text.txt"), sdo.encodingFormat.term.uri
     ) == Literal("text/plain")
 
-
-def test_bioschemas():
-    crate = make_test_crate(recursive=False)
-    crate.add_entity(
-        bioschemas_drafts.LabProtocol("#some_protocol"),
-        sdo.name(sdo.Text("Some Protocol")),
-    )
-    crate.validate()
-    crate_json = json.loads(crate.compile())
-    for entity in crate_json["@graph"]:
-        assert "http" not in entity["@type"], "All terms should be shortened"
-
-    assert crate_json["@context"] == [
-        "https://w3id.org/ro/crate/1.1/context",
-        {"LabProtocol": str(bioschemas_drafts.LabProtocol.term.uri)},
-    ], "Only the terms that are used in the crate should be in the context"
-
-
-def test_bnode():
-    crate = make_test_crate(recursive=False)
-    crate.add_entity(bioschemas_drafts.LabProtocol(BNode()))
-    assert any(isinstance(id, BNode) for id in crate.graph.subjects())
-
-
-def test_multi_type():
-    class CustomType(rdfs.Class):
-        term = RdfTerm("https://example.org/SomeOtherType")
-
-    with tempfile.TemporaryDirectory() as tmpdir:
-        crate = AttachedCrate(tmpdir)
-        crate.add_entity(
-            bioschemas_drafts.LabProtocol("#multi_type_entity"),
-            CustomType.to_type_property(),
-        )
-
-        # We never want to redefine rdf:type
-        assert str(rdf.type.term.uri) not in crate.context.to_dict().values()
-        # But we should have the custom type in the context
-        assert "https://example.org/SomeOtherType" in crate.context.to_dict().values()
-
-
-def test_redefine_term(empty_crate: AttachedCrate):
-    with pytest.raises(ValueError):
-        empty_crate.register_terms(
-            [
-                RdfTerm("https://example.org/Thing", "Thing"),
-            ]
-        )
-
-
-def test_add_metadata_term(empty_crate: AttachedCrate):
-    """
-    Test that terms used in `add_metadata` are registered in the context.
-    """
-    empty_crate.add_entity(
-        sdo.Thing("#thing"),
-        dc.valid(rdfs.Literal(Literal(True))),
-    )
-    # If valid can expand, it means it was correctly registered
-    assert empty_crate.context.expand("valid") == "http://purl.org/dc/terms/valid"
-
-
-def test_reverse_prop(empty_crate: AttachedCrate):
-    """
-    Test that we can add a reverse property to an entity.
-    """
-    empty_crate.add_entity(
-        sdo.Thing("#thing"), sdo.hasPart.reverse(empty_crate.root_data_entity)
-    )
-    assert (
-        empty_crate.root_data_entity.id,
-        sdo.hasPart.term.uri,
-        URIRef("#thing"),
-    ) in empty_crate.graph
-
-
-def test_adhoc_term(empty_crate: AttachedCrate):
-    """
-    Test that we can define terms on the fly and use them in properties.
-    """
-    thing = empty_crate.add_entity(
-        sdo.Thing("#thing"),
-        RdfProperty.adhoc(
-            term=RdfTerm("https://example.org/myProp"), object=sdo.Text("value")
-        ),
-    )
-    assert (
-        thing.id,
-        URIRef("https://example.org/myProp"),
-        Literal("value"),
-    ) in empty_crate.graph
-
-
-def test_cls_with_term_label(empty_crate: AttachedCrate):
-    """
-    Test that we can use two types with the same term label, by renaming one of them.
-    """
-    thing = empty_crate.add_entity(
-        sdo.Class("#thing"), owl.Class.with_term_label("owlClass").to_type_property()
-    )
-    assert (thing.id, RDF.type, sdo.Class.term.uri) in empty_crate.graph
-    assert (thing.id, RDF.type, owl.Class.term.uri) in empty_crate.graph
-
-
-def test_prop_with_term_label(empty_crate: AttachedCrate):
-    """
-    Test that we can use two properties with the same term label, by renaming one of them.
-    """
-
-    class exHasPart(RdfProperty):
-        term = RdfTerm("https://example.org/hasPart", "hasPart")
-
-    part = sdo.CreativeWork("https://example.org/part")
-
-    with pytest.raises(ValueError):
-        empty_crate.add_entity(
-            sdo.Thing("#thing"),
-            sdo.hasPart(part),
-            exHasPart(part),
-        )
-
-    empty_crate.add_entity(
-        sdo.Thing("#thing"),
-        sdo.hasPart(part),
-        exHasPart.with_term_label("exHasPart")(part),
-    )
-
-
-def test_adhoc_class(empty_crate: AttachedCrate):
-    """
-    Test that we can use two types with the same term label, by renaming one of them.
-    """
-    ExampleClass = RdfClass.adhoc(RdfTerm("https://example.org/thing"))
-    empty_crate.add_entity(
-        ExampleClass("#thing"),
-    )
-    assert (URIRef("#thing"), RDF.type, ExampleClass.term.uri) in empty_crate.graph
 
 
 def test_spaces_in_path(empty_crate: AttachedCrate):